Remove claim that source is C99.
Currently the code was compiled with -std=c99, promising
the it is C99 compatible source. That's not entirely true, since
we use readlink(), which doesn't play in the C99 field.

An alternative would be -std=gnu99, but since the code makes nowhere
any attempts to force this, i remove this option entirely.